requirements.txt
beautifulsoup4==4.12.3
discord.py==2.4.0
PyMuPDF==1.22.1
pandas==2.2.2
python-dotenv==1.0.1
pytube==15.0.0
openai-whisper
tiktoken==0.3.3
openai==0.27.4
